changes to health layout, reorganizing lottie animations folders, work on email
