+++ src/components/HealthSlider.jsx
@@ -1,4 +1,3 @@
-import { FontAwesomeIcon } from "@fortawesome/react-fontawesome";
 import React, { useState } from "react";
 
 export const HealthSlider = () => {
