+++ src/components/animations/FounderAnimation.jsx
@@ -1,6 +1,6 @@
 import React from "react";
 import Lottie from "lottie-react";
-import founderAnimation from "../assets/founder.json";
+import founderAnimation from "../../assets/founder.json";
 
 export const FounderAnimation = () => {
   return (
