+++ src/components/animations/SupportAnimation.jsx
@@ -1,5 +1,5 @@
 import Lottie from "lottie-react";
-import animationData from "../assets/support.json";
+import animationData from "../../assets/support.json";
 
 export const SupportAnimation = () => {
   return (
