+++ src/css/HelpCenter.css
@@ -102,8 +102,6 @@ Contact Section
 =====================================
 */
 
-/* ContactForm.css */
-
 .contact-form {
   display: flex;
   flex-direction: column;
